#6 Viewファイルの整理＋リファクタリング
