+++ monsterCardBattle/Date/MonsterData.swift
@@ -1,0 +1,16 @@
+//
+//  MonsterData.swift
+//  monsterCardBattle
+//
+//  Created by 江藤小夏 on 2025/09/18.
+//
+import SwiftUI
+
+// プレイヤーモンスターのモデル
+struct MonsterData: Identifiable {
+    let id = UUID()
+    let name: String
+    let atk: Int
+    let df: Int
+    let borderColor: Color
+}
