+++ monsterCardBattle/Shape/CutTopLeftShape.swift
